Replaced double sin()/cos() with float sinf()/cosf()

+++ src/badguy/flame.cpp
@@ -64,8 +64,8 @@ Flame::active_update(float elapsed_time)
 {
   angle = fmodf(angle + elapsed_time * speed, static_cast<float>(2*M_PI));
   if (!Editor::is_active()) {
-    Vector newpos(start_position.x + cos(angle) * radius,
-                  start_position.y + sin(angle) * radius);
+    Vector newpos(start_position.x + cosf(angle) * radius,
+                  start_position.y + sinf(angle) * radius);
     movement = newpos - get_pos();
     sound_source->set_position(get_pos());
   }
